requirements.txt
aiohttp==3.9.5
langchain==0.2.3
langchain_chroma==0.1.1
langchain_community==0.2.4
langchain_core==0.2.5
langchain_openai==0.1.8
python-dotenv==1.0.1
Requests==2.32.3
serpapi==0.1.5
google-search-results==2.4.2
streamlit==1.35.0
xmltodict==0.13.0